+++ driver-core/CHANGELOG.rst
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+2.0.10.1:
+---------
+
+- [improvement] Use Netty's pooled ByteBufAllocator by default (JAVA-756)
+
+
 2.0.10:
 -------